requirements.txt
pandas
scikit-learn
joblib
pyyaml
fastapi
uvicorn